Use an action queue/set for each push! to ensure actions aren't repeated and are performed in the correct order.
Instead of performing actions attached to a node as soon as a new value is sent to it, add the actions to an ordered set `action_queue` for the current `push!` message, keyed (de-duplicated) using the recipient node. This ensures that actions that would previously have been performed twice, or performed once before all the inputs to the action had updated (and ignored later after they had updated due to the `timestep` mechanism), will be performed once and at the correct time, i.e. after all their inputs have updated for this `push!` message.

Essentially causes the Signal graph to be traversed breadth first each push!, rather than the previous depth first.